Move policy engine models into a common location so other services can utilize them.
