+++ anchore_engine/services/policy_engine/api/controllers/distro_mappings.py
@@ -13,7 +13,7 @@ from anchore_engine.db import (
     get_thread_scoped_session as get_session,
     DistroMapping as DbDistroMapping,
 )
-from anchore_engine.services.policy_engine.api.models import DistroMapping
+from anchore_engine.util.models import DistroMapping
 from sqlalchemy.exc import IntegrityError
 from anchore_engine.apis.authorization import get_authorizer, INTERNAL_SERVICE_ALLOWED
