+++ anchore_engine/services/policy_engine/api/controllers/feeds.py
@@ -8,7 +8,6 @@ from anchore_engine.apis.exceptions import (
     AnchoreApiError,
     BadRequest,
     ConflictingRequest,
-    InternalError,
     ResourceNotFound,
 )
 from anchore_engine.clients.services.simplequeue import (
@@ -19,7 +18,7 @@ from anchore_engine.common.errors import AnchoreError
 from anchore_engine.common.helpers import make_response_error
 from anchore_engine.db import FeedGroupMetadata as DbFeedGroupMetadata
 from anchore_engine.db import FeedMetadata as DbFeedMetadata
-from anchore_engine.services.policy_engine.api.models import (
+from anchore_engine.util.models import (
     FeedGroupMetadata,
     FeedMetadata,
 )
