+++ anchore_engine/services/policy_engine/api/controllers/synchronous_operations.py
@@ -34,7 +34,7 @@ from anchore_engine.db import (
 )
 
 # API models
-from anchore_engine.services.policy_engine.api.models import (
+from anchore_engine.util.models import (
     Image as ImageMsg,
     PolicyEvaluationProblem,
     PolicyEvaluation,
