+++ anchore_engine/services/policy_engine/engine/vulns/mappers.py
@@ -4,7 +4,7 @@ from collections import defaultdict
 from typing import List, Dict
 
 from anchore_engine.db import Image, ImageCpe, ImagePackage
-from anchore_engine.services.policy_engine.api.models import (
+from anchore_engine.util.models import (
     VulnerabilityMatch,
     Artifact,
     Vulnerability,
