+++ anchore_engine/util/models.py
