+++ tests/unit/anchore_engine/services/policy_engine/api/test_models.py
@@ -1,5 +1,5 @@
 import datetime
-from anchore_engine.services.policy_engine.api.models import (
+from anchore_engine.util.models import (
     FeedMetadata,
     FeedGroupMetadata,
     Image,
