+++ tests/unit/anchore_engine/services/policy_engine/engine/vulnerabilities/test_deduper.py
@@ -1,6 +1,6 @@
 import pytest
 
-from anchore_engine.services.policy_engine.api.models import (
+from anchore_engine.util.models import (
     VulnerabilityMatch,
     Vulnerability,
     Artifact,
